+++ tetst/class/basic/MyRect.h
@@ -11,14 +11,14 @@
 
 @interface MyRect : NSObject{
     MyPoint* pos;
-    int length;
-    int width;
+    float length;
+    float width;
 }
-- (void)move:(int)x:(int)y:(int)w;
+- (void)move:(float)x:(float)y:(float)w;
 
 @property (readwrite, nonatomic) MyPoint* pos;
-@property (readwrite, nonatomic) int length;
-@property (readwrite, nonatomic) int width;
+@property (readwrite, nonatomic) float length;
+@property (readwrite, nonatomic) float width;
 
 - (CGMutablePathRef)generatePath;
